Allow dynamic page titles with a sensible default
Pages should have unique page titles and this gives us a way to set the
page title for any view by setting the value of page_title in the view
file.

If none is set we will use the service name from the locales file
which this commit also adds.

Wherever the service name appears we should us t(:service_name) so
we can change the service name at any point from a single place.