(chore) Remove Rubocop from the application
In preparation for installing the `standardrb` gem that will look after
the linting configuration for us so we have one less thing to manage.

We were using `dxw-utils` as an effort to have such a 'default'
configuration. However it hasn't seen many updates and still requires
managing, using a community driven config feels more sensible.