(chore) Add standard gem for linting our Ruby code
Linting code is a good thing, having to configure a linter leads to
bikeshedding. Standard gives us a predefined, community agreed
configuration so we don't have to worry about it.

This could have an impact on running Rubocop in your editor, if you have
issues please look at this section of the readme for standard:

https://github.com/testdouble/standard#how-do-i-run-standard-in-my-editor