(chore) Add standard to the default rake task
To enable us to run `standard` as part of our CI build process.

It's worth noting that by default if standard doesn't find any errors it
won't print anything to stdout. It will only print on errors. We may
decide to change the default formatter at a later date but the defaults
should be fine for now.