Avoid SchoolDataImporter memmory bloat
CSV#parse creates the entire CSV data structure in memory in order to
parse it, which make the code very expensive to run for large csv files
and was making the rake task exceed the memory limit of our current
Heroku configuration.

CSV#foreach allows the data to be streamed from a file instead, which is
much more memory efficient, but requires that we be more explicit about
creating a Tempfile when downloading the data. This is because URI.parse
().open() tries to be clever and returns a `StringIO` object if the data
payload is small enough (and CSV#foreach doesn't like that).

There may be a negligible increase in the time this will take but the
memory saving is large and the file is on disk already as we had to
download it: our testing on the school data in April 2019 we went from
1034.17 MB to 80.88 MB.

Prior art from:

https://dalibornasevic.com/posts/68-processing-large-csv-files-with-ruby