Rework routing to use #show for claim sequence
Rather than having an action per-page, we will route pages in the claim
sequence to claims#show and show the user the appropriate page from the
sequence. Whilst we only have two pages in the sequence (post-consent),
this can be kept super simple and hard-codes the redirect path on the
update action to the second page in the two page sequence. Expecting
more modelling once more pages in the sequence are added.