User can search for schools
So that a user can tell us which school they were employed at.

Note that we've added some test schools using Rails fixtures. We've made the
decision that a small number of carefully curated fixtures for things that are
fundamental to the service (e.g. a eligible school; an ineligible school) will
make writing and maintaining the specs easier, and also running them faster.

Co-authored-by: Tekin Suleyman <tekin@tekin.co.uk>