User can select and save a school
The user can select the correct school from the list of schools that match
their search criteria and the selection is saved to their claim.

We decided to use radio buttons to make the school selection for this
iteration as it was the clearest way for a user to select a matching school.

Fixes a typo on the fixture for 'Penistone Grammar School'.