Limit the school search term to four characters minimum
If the user enters a short search term we may return a very large number
of results. By checking the length of their search term we can limit the
number of school names in the search results.

The ambition is to have a JavaScript auto-complete for users who can utilise
it. As we are progressively enhancing the service this first version calls a
method named 'perform_non_js_school_search' to be clear.