Fix sequential migration dependency error
When running `rails db:migrate` I got the error
`PG::UndefinedFunction: ERROR:  function gen_random_uuid() does not exist`

After inspection it looks like the the `CreateTslrClaims` migration
happens before we enable the pgcrypto extension causing the error.

The fix puts the `CreateTslrClaims` migration in the correct place that
it occurred sequentially. As it is the most recent migration we need to
point the schema to it as well.