Determine if a given school meets the TSLR eligibility criteria
The list has been compiled from the eligibility criteria here:
https://www.gov.uk/guidance/teachers-student-loan-reimbursement-guidance-for-teachers-and-schools

To be eligible for TSLR payments a teacher must've taught in a school
that meets the following criteria:

- State funded school in England
- Local Authority maintained secondary (including middle deemed
  secondary)
- Local Authority maintained or non-maintained special schools
- Secondary phase academy/free school
- In one of the 25 participating local authorities

Using the data from the Get Information About Schools database and after
talking with the Financial Incentives policy team we can make these
assertions about school eligibility:

- State funded school means any school that isn't independent, a
  university, a welsh school or 'Other'
- Local Authority maintained, Colleges, Academies, Free schools and
  special schools are only eligible if they have a phase of secondary or
  middle deemed secondary
- All special schools are eligible except independent special schools.
  Note: The only way to filter out independent special schools is with
  the 'school type' field (they will have a 'school type group' of
  'special_schools')

We may also want to capture the date a school closed and also add a
check to see if the school closed before the year the teacher is
claiming for, however this wasn't part of the official criteria so I
haven't included it in this commit. This means that it will be possible
for a teacher to select a school that otherwise would be eligible but
closed last decade.