A teacher can specify their current employment
Teachers must still be teaching to be eligible for TSLR so we ask them
if they are still employed.

The next stage of the journey will optionally ask them their current
school if they say they are teaching at a different school, or tell them
they are ineligible if they say they are no longer employed, both coming
in future commits.

Co-authored-by: mec <mec.rawlings@gmail.com>