Get current school if no longer at claim school
A teacher may no longer be teaching at the school they were originally
teaching at during the claim period. If they are teaching at a different
school, we want to ask them to confirm that school in case we need to
contact the school to confirm any details of their claim.

Co-authored-by: mec <mec.rawlings@gmail.com>