Admin page is only accessible to users with authenticated session
Users can create an authenticated session. We currently have no
authentication in place, that will come later on this branch. We protect
the admin controller with a check to see if the authenticated session
exists, if not redirect to the sign in page.

The admin page is painfully simple at the moment but the aim is just to
create a "protected" page. The actual content and functionality will
come later.