Ensure the current_school value is never nil
When a teacher is still employed at the same school as the on which they are
making a claim for, their employment_status is set to claim_school and they
skip the question that sets the current_school value.

We still want to record their current school to keep the database values
accurate, even though it is the same as the claim_school value.

NOTE: We came across what looks like a bug (see link) in the Bullet gem so we
white-listed the unused_eager_loading on the association used in the testing
environment.

https://github.com/flyerhzm/bullet/issues/435

Co-authored-by: Tekin Suleyman <tekin@tekin.co.uk>