Remove the Rails error class from all form fields
We noticed that when Rails adds the 'field_with_error' css class to the
label and input elements in a form, the GOVUK styles can be broken due to the change in the mark-up.

The 'filed_error_proc' simply returns the html_tag wrapped in a span with the
class name so we just override that with no wrapping.

We expect future work to render the correct GOVUK errors on elements to render
this work redundant.