Use fixtures for TSLR eligible local authorities
The list of eligible local authorities is determined using a constant
set of local authority codes defined by
Tslr::SchoolEligibility::ELIGIBLE_LOCAL_AUTHORITY_CODES. This makes it
difficult to use factories to define eligible/ineligible local
authorities because of the uniqueness constraint on the code attribute.
Instead we introduce fixtures for eligible and ineligible local
authorities.

This also changes the sequence on the local authority so that the
default factory won't cause clashes with the real local authority codes
(these currently only go up to 938).