Choosing an ineligible claim school ends journey
For now we're going for shameless green and keeping this simple. Some
form for abstraction for claim eligibility will likely emerge once we
start implementing the other ineligibility checks, i.e. no longer
teaching and not teaching 50% of the time. We also keep the content on
the page static for now. Varying the content based on the reason for
ineligibility will be best done when we have more than a single reason
for ineligibility.