Capture full name from claimant
This (along with the address and date of birth) will eventually come to
us via the payload we will get back from a GOV.UK Verify verification
step. Until we integrate with Verify we'll simply ask for the
information using the standard GDS design pattern.

Limiting the database field to 200 characters, which should be more
than enough for any name, whilst minimising the impact on the database
of malicious actors posting large payloads.