Capture address of claimant
Capture the address fields directly from the user for now until we do
the GOV.UK Verify integration.

The format of the address (several lines followed by a "postcode" field)
was chosen to map to the format we expect addresses to come back from
Verify in[1].

The form is following the pattern in the GOV.UK design system for
capturing addresses as multiple text inputs[2]

Only the first line, postcode and line that corresponds with the "Town
or city" field are required.

Again we limit the lengths of these fields to minimise impact on the
database of mailicous actors posting large payloads.

[1] https://github.com/alphagov/verify-service-provider/blob/master/docs/api/verify-service-provider-api.swagger.yml
[2] https://design-system.service.gov.uk/patterns/addresses/#multiple-text-inputs