Add simple date of birth capture to the journey
I say simple... we have to do some rails form gymnastics so that the
user can provide their date of birth as separate fields.

The validation here is a generic one scoped to the entire date, i.e. it
won't validate the individual components of the date as described in the
GOV.UK design system. Adding individual field validations would be
possible, but I've decided to skip it because a) it's a fair chunk of
work because the way we map the fields to the date doesn't line up
neatly with how ActiveRecord errors works, and b)  we plan to replace
this question with GOV.UK Verify integration, which will give us the
date of birth without the claimant having to explicitly provide it.