Normalise teacher reference number by removing non-digits
We want to allow a teacher to input their teacher reference number in any
way they want, but store it in a consistent format.

We settled on storing only the digits in the TRN to simplify the normalisation.

We initially used a setter for normalising the TRN however we found that
when we added minimum length validation when the user received an error
the value they entered would be changed to the normalised value which
might be alarming. Instead we only normalise the value when we save the
record.

Co-authored-by: robbiepaul <robbiepaul@me.com>