Ensure teacher reference number is 7 digits long
At the moment we know that TRNs should be 7 digits, therefore if a
teacher enters a number shorter or longer than that we should show an
error.

TRNs can contain non-digit characters so we strip those out before we do
the validation.

We always want to validate the TRN if it exists.

Co-authored-by: mec <mec.rawlings@gmail.com>