Time out claim session after 30 minutes
To help keep users’ information private we end the claim session after
a fixed period of inactivity. We've picked 30 minutes for the time being
but this may change in the future.

Note that we don't do anything with the incomplete claim data that will
still be stored in the database just yet. At some point we want to
ensure that this data is not retained, and perhaps deleted when the
session is timed out, but until we have a way to distinguish between a
fully completed claim and a partial one, we risk deleting completed
claims by mistake. Capturing this elsewhere for a future story.