Validate National Insurance number format
We're being quite loose in our validation of NI at the moment, following
the guidance on the GOV.UK design system. We know that some of the two
letter prefixes cannot be certain combinations, we think this is ok to
leave off for now as it is a first iteration. If we put too many rules
on the validation we could end up with edge case issues.

We only validate the normalised value so we can be more explicit with
our pattern matching.

Co-authored-by: robbiepaul <robbiepaul@me.com>