Claimants can say if they have taught eligible subjects
The content on this page comes from the TSLR eligibility checker from
alpha, we may want to improve it at a later date. We have simplified
the options slightly to just a "Yes" or "No" answer, this meant we need
to be a bit more explicit about the question we're asking.