Add length validation to contact fields
We have a database length limit on the following fields but no
validation to check it when submitting:

- name (200)
- address_line_1 (100)
- address_line_2 (100)
- address_line_3 (100)
- postcode (11)

Users trying to submit values longer than this would cause an error in our app,
so this adds a length validators to stop the problem.

This commit also:

- Adds some white space around the validation definitions as
they are getting large and hard to read.

- Re-orders the specs as it was getting hard to navigate them. We have tried
to show the different validations that are performed in different validation
contexts.