Determine if a school is eligible for MPTR
The official criteria can be found here:
https://www.gov.uk/guidance/apply-for-mathematics-and-physics-teacher-retention-payments#eligibility-criteria

We used the same criteria as the TSLR claims as only the area a school
resides in is different.

To be eligible for mathematics and physics teacher retention payments a
teacher must've taught in a school that meets the following criteria:

- State funded school in England
- Local Authority maintained secondary (including middle deemed
  secondary)
- Local Authority maintained or non-maintained special schools
- Secondary phase academy/free school
- In one of the 42 listed local authority districts

The policy states 'Fenland and East Cambridgeshire' which is two separate
districts in GIAS so we end up with 43 local authority districts in code.

The policy also lists 'Kingston upon Hull' which is 'Kingston upon Hull', city
of' in GIAS.