Enable https for development environment
DfE Sign In requires both parties to use SSL, this means that we need to
use self-signed certificates for our development server.

Add a `Procfile.dev` file for starting the development server with the
self signed certificate. This should help developers in not having to
remember the exact flag and arguments needed.

Foreman doesn't load in `.env.local` so we need to use `.env` to load in
our environment variables. This has a side effect of needing extra env
files for development (without DfE Sign In) and test.