Allow a user to provide their bank details
In order to pay a teacher their TSLR payment, we need to capture their bank
account details.

At this stage we are not collecting a separate 'account holder name', we hope
this information will come from another source.

This commit does not include normalising the values a user provides, this will
come later.

The data we store may need to be encrypted as it is sensitive, this will also
come later.