Add custom validation to bank details
To simplify the entry of a users bank account details we strip common
characters and then validate.

- Normalisation removes white space and “-”
- Validation checks the length and only digits were used

Both values can be nil in any validation context other than 'bank-details',
when they are required.

If a value fails the validation, the user is shown what they entered rather
than the normalised value to aid them to fix a problem.

Re-factored the TSLRClaim spec, removed the "when saving a TslrClaim" context
as it wasn't making sense and added the "when saving" when needed.