Pull admin and auth controllers in to an admin namespace
Both controllers extend from a new base admin controller. Doing this
clearly separates the admin and authentication logic from the rest of
service and will allow us to lock down the "/admin/*" endpoints in the
future.

This change will require us to change the callback URL that is
registered with DfE Sign In.

One annoying duplication is repeating the `prefix_path` in the OmniAuth
spec helper. It seems that the config options are getting overriden in
the specs causing the paths to be incorrect.