Add headless chrome for testing JavaScript features
OmniAuth needs to be told to `redirect_to_failure` in the specs
otherwise it will raise an exception.

We need to whitelist the chromedriver download URL so that the
`webdrivers` gem can download the executable.

Installed `launchy` gem so that we can use the Capybara helper
`save_and_open_screenshot`

- Chromedriver couldn't find the executable so we have to explicitly
  tell it to use the one in the environment variables.
- After this the browser was becoming unresponsive, because we are now
  using the bin and not the shim we have to tell chrome that, with the
  arguments `disable-dev-shm-usage no-sandbox`