Redirect to answer check page if editing
I've also had to add another Bullet whitelist, as if I remove the need for an N+1 query for this instance, I get lots of warnings about unneeded includes elsewhere. It's not a huge deal to do an additional DB query in this case.

There were also some issues in the validations, where the bank sort code was not being validated at the submit stage, so I've fixed this too