TSLR claim views use form helper
When searching for a school without a search term, the 'school search' error is added to the school_search attribute rather than 'base'. This means the
error_tag method can pick it up.

Changes to the views meant "School name" could not locate the correct element
in feature tests, using :school_seach fixes this.