Set up a first-pass web app on Azure
We're building from @pezholio's private DockerHub repo for now.

We're building in a new resource group, not the ones that appear to have
been created for us. This should probably change.

We include parameters for the development environment. These should be
changed and put into a secret store and ignored from Git.

All of the deployment names should be versioned to support rollbacks,
but for now, they all overwrite each other.