Validate that `current_school` is set on submit
It's set by the `update_current_school` callback if the user selects
same school, otherwise the user should be setting it manually.

We now run `update_current_school` on validation as well as saving. This
is fine to do as the value of it is invisible to the user. The other
callbacks only run on save so we show the user the values they entered
if there's an error with them, not a normalized version of them.