Fix intermittent failing JavaScript redirect spec
It's possible for the redirect to take longer than a second, to fix this
we can leverage Capybara's `synchronize` helper which will wait until
the block has completed up until `default_max_wait_time`.

Extract the functionality out to a helper so that it reads better in
the spec.

We also need to increase the `default_max_wait_time` to 5 from the
default of 2.