The link used in a GOVUK error summary can be overridden
The link in the error summary was set to an anchor with the combination of the
object model name and the attribute name, this works in most cases except for
forms using radio buttons, check boxes or other multi-element forms that
represent a single attribute (e.g. date of birth).

In these cases, the first element should be linked. By allowing the id to be
overridden when needed the required behaviour can be maintained.

This commits only changes the date of birth view to use it, there may be others
that will be updated in a future commit

Read more about the behaviour:

https://design-system.service.gov.uk/components/error-summary/