School search error summaries links
The claim school and current school views use radio buttons on the second step.
As per the GOVUK guidance, the error summary should link to the first radio
button when there are validation errors.

Using the GovukFormHelper the error summary link is overridden to be the first
radio buttons id, which is the first school in the list of results.

Because the school search views use two forms, the simplest way to show the
correct error summary was to include one in each part of the view.