Only reset inferred current school when persisted
The build broke because the before save callback was changed to a before
validation callback. Skipping and setting callbacks in the factories
makes the code brittle. Instead of doing that, this reworks the callback
so that it only resets the current school when the employment status
changes on a persisted record, as it's only on persisted records
(created as part of the claim journey) that we expect to automatically
set/reset the current school based on the answer to the employment
question.