Address CVE-2015-9284
There is a known vulnerability in OmniAuth that could potentially be
used in combination with a vulnerable OAuth providor to gain a signed-in
user's session. Whilst DfE Sign-in uses OpenID, not OAuth, the substance
of this patch (adding CSRF protection to the "Sign in" link on the
service) is still worth doing.

For more background see:

https://github.com/DFE-Digital/dfe-teachers-payment-service/network/alert/Gemfile.lock/omniauth/open
and https://github.com/omniauth/omniauth/pull/809
which leads to https://github.com/cookpad/omniauth-rails_csrf_protection