Add eligibility confirmation page
Once a user has answered enough questions for the service to confirm they
are eligible to make a claim we show them them a confirmation page.

It was felt important the GOV.UK Verify content and button was included but
disabled to remain close to the real user journey. Once integrated, skipping
GOV.UK Verify will not be possible.