Add controller for generating Verify auth request
In the spirit of keeping things RESTful, and following the gudiance in
the [verify documentation], this controller action will render a form
that auto-submits (with JavaScript eventually) a GOV.UK Verify
authentication request to Verify. For now we're hardcoding nonsense
values into the form until such a point as we have an interface to speak
to the VSP and get real values back.

[verify documentation]: https://www.docs.verify.service.gov.uk/get-started/set-up-successful-verification-journey/