Introduce Verify::FakeSSO to simulate verification
In our tests, we want to be able to simulate the GOV.UK Verify process,
sending the user off to the Verify SSO for them to attempt verification.
For actual real-life Verify, a form is POSTed back to our callback URL
with a SAMLResponse. This Rack app simulates that process, rendering a
form that can be submitted in a feature spec to send a fake response to
our callback URL.

In the future it will be able to return different forms for the
different types of responses we need to handle (i.e. NO_AUTHENTICATION
and AUTHENTICATION_FAILURE).

Co-authored-by: mec <mec.rawlings@gmail.com>