Consistently ask questions
Previously we had a combination of questions and statements where we
required users to provide answers. After some discussion we found some
guidance in the GOV Service Manual that said either way was fine but you
need to be consistent.

This should address where we ask and review the questions.