Move webdrivers to development group
The alpine implementation of webdrivers seems broken. It downloads, then
removes the binary straight away then complains that its missing.

We don't need to download chromedriver through webdrivers gem because we
add it to our Docker image.

Developers will still want the ease of having this helper download
chromedriver so we can move it to the development group.