Run rake inside Docker on Azure Pipelines
Setting up our app for running tests and linting is quite involved and
would require us to keep a separate YAML file up-to-date with all the
required steps.

Instead we can use docker-compose to spin up our app container and run
the specs inside. It requires that we modify the Dockerfile to ensure
that we're installing the correct dependencies and files for testing.