Build the Docker image for deploying
We have to explicitly tell the pipeline to use a variable group.

Secret variables need to be defined as an `env` before usage so that
they can't be printed out in the build output.

They are replaced by '***'

We make sure that we are building the image with the target of `web` so
that only production assets are added.