Overrride the VSP test identity dataset for LOA2
The VSP returns a [stubbed identity defined as JSON] in development
mode. The version built-in to the VSP returns an LOA 1 response, but our
application needs to request and verify at level 2. We can override the
default test identity at boot-time by passing in some JSON data.

In the future we'll roll all this up into a nicely packaged bin script
so that the VSP can be started up using the foreman Procfile.dev.

[stubbed identity defined as JSON]: https://github.com/alphagov/verify-service-provider/blob/master/src/main/resources/default-test-identity-dataset.json