Add GOV.UK Verify callback endpoint
This implements the happy path of a successful IDENTITY_VERIFIED
response from GOV.UK Verify.

A user completing the GOV.UK Verify process will trigger a POST back to
our service with the results in a SAMLResponse parameter. We need to
pass this back to the VSP, along with the request ID and the level of
assurance our application is configured to require.

CSRF protection is disabled for this endpoint as the POST will be coming
from Verify, not from our service.

Limitations at this stage:

  - The first value for attributes returned as arrays (e.g.
  "firstNames") are stored without actually checking if it's been
  returned as "verified"
  - Only IDENTITY_VERIFIED responses are handled
  - Non-UK addresses are not handled
  - EU identities are not handled