Use delayed_cron_job to run jobs on a schedule
This adds a new db:schedule_jobs rake task to schedule jobs. This task
runs after db:migrate and db:schema:load to automate populating the
queue with the scheduled jobs.

CronJob supports setting jobs to perform immediately when scheduled to
support running task now and on a schedule, such as for the school data
import.