Change half hours question
The second part of the 50% question was shown to be ambigious in user
testing. Users were unsure what constited 50% of their teaching time,
50% of hours spent teaching or 50% of contracted hours.

The question was rephrased and hint text expanded upon to make this
clearer for users.

The ineligible page header was also edited to make it clear users
were not eligible for the service.