Begin the student loan plan micro journey
So we can calculate the tax a user will pay on their student loan repayment
we need to know which student loan 'plan' they are on.

This is not information a user will be able to find out simply so
instead we ask a series of questions that determine the the correct plan or
plans a user is on.

This commit introduces the first question in that series.