Amend the student loan amount questions
UR was showing that people were likely to just guess rather than enter
the precise amount with the expectation that DfE or HMRC would pick up
the discrepancy and correct it.

The text was amended to encourage users to provide some more accurate
information and will be tested in a future sprint.