Add the student loan start date question
In order to determine the student loan plan a user is on we also need to
know when they started their course or courses.

Because a user could have completed multiple courses there are two
versions of this question. One for a single course and one for more than
one course.

The view was kept as simple as possible by switching on the answer to
the previous question, 'How many higher education courses have you
studied?'.

All of the steps are now in-place to determine the student loan plan,
which is the next step.