Amend QTS question to use ITT
Users were struggling to answer the QTS question, they were unsure if
they answered Jun YEAR when they got QTS or Sept YEAR when they started
at the school (Newly Qualified Teacher).

We changed it to use Initial Teacher Training as there were no issues
with this in the maths and physics journey.

Amended CI to use the alias for the question so we don't have to keep
updating it.