Remove Verify feature flag
We intend to always use Verify in development so we can safely remove
the feature flag.

Changed the main claim journey to be JavaScript enabled as there's some
JavaScript as part of the Verify process and most people will have JS
enabled. We still have coverage for the non-JS version in the Verify
spec.