Move logic for updating a claim to separate model
The logic around updating a claim in the UI is about to get more
complicated with the addition of the student loan plan questions, and
also the forthcoming changes to the 50% question, particularly when the
user goes back to "change their answers". In preparation for that we're
extracting a separate object to manage the changes users make to a
claim.

Note that because we're now passing `claim_params` into the new object
even when submitting the claim in the "check-your-answers" context, we
need to ensure it doesn't explode when there aren't any `:tslr_claim`
parameters present, hence the change from `require` to `fetch`.

Co-authored-by: mec <mec.rawlings@gmail.com>