Update presence error messages for name, address and date of birth
The presence error messages will no longer be seen by users as we're
getting the name, address and date of birth back from Verify.

We still need them for the cases when the records are updated manually
by the team.

This could present an issue if not all the required data is returned
from Verify, the user would not be able to recover as it stands. We feel
like this is enough of an edge case not consider at the moment and it
will need some design thinking about how this re-verification process
will work. We'll create a card in the backlog for it.