Move logic for auto-setting current_school
Tying this sort of logic to the model as ActiveRecord callbacks can
result in pain because it is difficult to bypass when we need to. We're
also about to introduce a bunch of similar logic for other questions
(student loan plan; 50% question), which will only make the problem
worse. By putting the behaviour in the ClaimUpdate model we ensure that
dependent/inferred attributes are only modified/set for changes
performed using the user interface.