Reset employment_status when claim_school changes
If a user changes the claim_school, then the answer they previously gave
for employment_status will no longer make any sense, so we should reset
it (which also triggers a reset of the claim_school).

The upshot if this is we can simplify PageSequence as it no longer needs
special-case handling for making changes to the dependent schools
answers in the as the claim will remain unsubmittable until the newly
reset questions are corrected. This has the added benefit of not forcing
the user to re-answer the still-teaching and current-school questions if
they don't actually change the claim school (those questions can be
changed separately by clicking the appropriate link on the confirmation
screen).

This is going to come in handy as we will be following the same pattern
for the student loan plan questions.