Add the how many student loan courses question
In order to determine the student loan plan for a user who received
their loan in England or Wales we have to find out how many courses they
have taken.

If the user received their student loan in either Scotland or Northern
Ireland this question is skipped as they can only be on Plan 1.

Co-authored-by: mec <mec.rawlings@gmail.com>