Allow state funded special schools where their phase isn't specified
Most special schools have a "Not Applicable" education phase, so this
adds support for those schools.

We use the school type, rather than the school type group for
determining if a school is a special school because e.g. free special
schools are grouped under free schools, and so would be left out. There
maye be other examples.