Change QTS award year to an enum
To be consistent with other pages in our journey and we also get the
inclusion validation for free.

It's worth noting that some of the typical enum helpers won't work. eg.
`claim.awarded_qualified_status_2013-2014?` isn't a valid method name
because it has a `-` in it.

The migration has both an `up` and `down` method incase anyone wanted to
rollback through the migrations.