Claim submittable when loan plan can be determined
Certain student loan-related questions are only answered under certain
circumstances, so their presence cannot be a requirement for submission.
Instead, we put the requirement on the student loan plan being
determinable. To do this, we introduce the "not applicable" option which
can be set once we have enough answers to determine when that is the
case.

In theory a user following the standard flow of the service should never
see the new validation message for student loan plan, but I've tried to
make it make as much sense as possible just in case.