Require explicit stub of successful auth attempts
This makes it such that the expected response from a DfE Sign-in auth
attempt is undefined by default, and has to be explicitly set before
each spec as required. Prior to this, it was set once on initialisation
in support/omniauth.rb and then became whatever subsequent specs set it
to, meaning it would have a random value depending on the order of the
specs being run.