Refactor the resetting of dependent answers
This commit introduces a data structure to define question/answer
dependencies and generic code to reset when appropriate.

There are two varieties of dependent answers:

  - an answer to a question where the question changes based on a
    previous answer, for example, the employment status question is
    based on the answer to the claim school question
  - an answer to a question where the question is conditionally asked
    based on the answer to a previous question, for example, we only ask
    the student loan country question if the user has a student loan

These dependent answers need to be reset whenever the user changes the
answer to the question it is dependent on so that the user is either
re-asked the question, or so it can be skipped by `PageSequence` because
it is no longer relevant.