Introduce model for StudentLoans::Eligibility
This will be the model where we store and determine the eligibility of a
claim against the student loans policy. It only contains attributes that
are specific to the Student Loans policy. This is the first step in
separating the policy-specific questions and eligibility checks from a
generic "Claim" model such that the codebase can be easily updated to
support claims for the Maths and Physics policy.

We'll migrate the behaviour for each individual attribute in separate
commits.

Co-authored-by: mec <mec.rawlings@gmail.com>