Step 1 migrating eligibility attributes: QTS year
First stage of moving all the data related to eligibility over to the
new StudentLoans::Eligibility model. We're going to migrate across one
attribute at a time to make the work easier to follow and review and
keep things green as we go.

By using nested attributes, the forms and flow can remain largely the
same, with the validations triggering on the eligibility, keeping the
user experience the same.

Note that whilst we do this step-by-step migration, a bunch of temporary
changes are necessary:

- We will delegate the reader for the attributes to the eligibility
whilst the migration is in progress, but remove them later, as the claim
shouldn't know about the individual eligibility criteria
- Same goes for ineligibility checks - they are delegated for now, until
we have moved everything across, at which point they can be removed from
the claim
- The other scoped validations need to be disabled (i.e. commented out)
from Eligibility until their attributes are also migrated over.

Co-authored-by: mec <mec.rawlings@gmail.com>