Use base64 encoded the private keys
Previously this base64 encoding was a manual step and it was one more
thing that needed documenting, now it's automated.

It happens as part of the script that requests the VSP certs.
When the private keys are created, we strip out the first and last lines
then remove all new line characters.

If we generated the key, we always want a new version of the base64
encoded version, regardless of whether we already had a base64 encoded
key or not.