Verify requests use SSL for https URLs
I'm not sure how to test this, WebMock doesn't give many options for
making assertions about requests.

This was causing requests to our integration VSP to be rejected as it's
running on https. We still want to be able to use the development VSP
so we simply check the port inferred from the URI. Any URIs with `https`
will have the port 443 inferred.