Don't raise on bullet warnings in test suite
Instead we will rely on the bullet logs.

This was causing us pain whilst working on the refactoring work to
extract a separate model for eligibility criteria because it was raising
circular warnings – if we didn't preload eligibility: [:claim_school],
etc when loading the claim in current_claim, it would complain that we
weren't eager-loading the associations, and if we did, it would complain
that they were being eager-loaded but not used. This is because not all
the pages where the current_claim is accessed also access the eligiblity
and/or the claim_school.

We could continue to add to the whitelist for these sorts of situations,
but then bullet starts to lose its value as we'll essentially end up
having to whitelist both n_plus_one_query and unused_eager_loading for
pretty much every association on claim and eligibility.