Avoid stubbing in-progress claim in feature spec
Going through the UI rather than stubbing a "submittable" claim makes
this spec more robust, as it is less impacted by changes to the
underlying data organisation. We were experiencing problems where we
had to update this spec every time we made a change to the way the
attributes were organised. This change means we'll have to make fewer
changes to the spec when the models underneath change.