Render a 404 for fields coming from Verify
Although we don't show the link to edit a field if it's come from Verify, there's always a risk that a nefarious user might try and access the field's edit page by guessing the url. This guards against this happening by returning a 404 if the field is showing as having come from Verify. I did consider being more technically correct and returning a 403, but in the sprit of not being kind to bad actors, I thought a 404 would be more appropriate.