Move schools-related attributes to Eligibility
This continues the work to separate the common claim-related data
(identity, payment, payroll, etc) from the policy-specific fields,
making it easier to support more than one type of policy (i.e. Maths &
Physics).

This commit focuses on the two schools associations and the closely
associated employment_status field. Because all three attributes are
closely related and interdependent they've had to be migrated at the
same time.

Note that no effort has been made to migrate the data from existing
claims as we're not yet live so any existing data can simply be deleted.
The best way to do this on the Azure environments is probably going to
be to drop the database and perform a release to force the rebuilding of
the database.

-> Temporary fixes and future work

- As with the preceding work that moved qts_award_year, a bunch of
  attributes are temporarily being delegated from the claim model to the
  eligibility model to keep these changes as small and focused as
  possible. These will need removing before Maths & Physics can be
  introduced as at that point, the claim model will need to be entirely
  generic and not know anything about the policy-specific eligibility
  fields.
- The ClaimUpdate model has needed to be updated with knowledge specific
  both to the claim model and the eligibility model. The code doesn't
  look great at the moment, but we know it will need adjusting at the
  point the new policy is introduced, and that feels like a better point
  to figure out the best way to model the behaviour for adjusting
  in-progress claims (and their eligibility) and their dependent
  attributes
- Same goes for many of the specs; many have had to be updated with
  student loans-specific attributes and knowledge, in some cases
  coupling them to the specifics of student loans. As above, we'll be in
  a better place to try and untangle these at the point the new policy
  is introduced.